Added: + Images for items + Decorator login_required + Json fix + xml endpoint + db_populate from json + csrf protection for delete added + readme updated

+++ run.py
@@ -1,11 +1,11 @@
 """
-catalog_functions.py: main functionality for Catalog App:
+run.py: main functionality for Catalog App:
     - authorization - Login/Logout
     - view categories
     - view/create/update/delete items
 """
 from flask import Flask, render_template, request, redirect, url_for, jsonify,\
-    flash, Response
+    flash
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from database_setup import Base, Category, Item
@@ -18,24 +18,25 @@ from flask import make_response
 import requests
 import random
 import string
-from xml.etree.ElementTree import Element, SubElement, tostring, dump
+from xml.etree.ElementTree import Element, SubElement, tostring
 from flask.ext.responses import xml_response
 from werkzeug import secure_filename
 import os
 import datetime
 from flask import send_from_directory
-import pdb
+from flask.ext.seasurf import SeaSurf
 
-#pdb.set_trace()
 
 UPLOAD_FOLDER = 'image/'
-ALLOWED_EXTENSIONS = set(['png','jpg','jpeg','gif'])
+ALLOWED_EXTENSIONS = ['png', 'jpg', 'jpeg', 'gif']
 CLIENT_ID = json.loads(open('client_secrets.json',
                             'r').read())['web']['client_id']
 APPLICATION_NAME = "CatalogApp"
 app = Flask(__name__)
+csrf = SeaSurf(app)
 app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 app.config['MAX_CONTENT_LENGTH'] = 2 * 1024 * 1024
+
 
 engine = create_engine('sqlite:///catalog.db')
 Base.metadata.bind = engine
@@ -44,27 +45,38 @@ DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
 
+@csrf.exempt
 @app.route('/catalog/image/<filename>')
 def uploaded_file(filename):
+    """
+    :param filename: name of uploaded file
+    :return: file
+    """
     return send_from_directory(app.config['UPLOAD_FOLDER'], filename)
 
 
-
-
-def xmlCategories(categories_xml,i):
-        category_xml = SubElement(categories_xml, 'category',
-                                  {
+def xml_categories(categories_xml, i):
+    """
+    :param categories_xml: parent element
+    :param i: current category
+    :return: generated xml for category i
+    """
+    category_xml = SubElement(categories_xml, 'category',
+                              {
                                       'id': str(i.id),
                                       'name': i.name,
                                       'description': i.description,
-                                   }
-                                  )
-        return category_xml
+                              })
+    return category_xml
 
 
-def xmlItems(items_xml,j):
-    SubElement(items_xml, 'item',
-                            {
+def xml_items(items_xml, j):
+    """
+    :param items_xml: parent element
+    :param j: current item
+    :return: generated xml for item j
+    """
+    SubElement(items_xml, 'item', {
                                 'id': str(j.id),
                                 'name': j.name,
                                 'description': j.description,
@@ -72,35 +84,48 @@ def xmlItems(items_xml,j):
                             })
 
 
+@csrf.exempt
 @app.route('/catalog.XML')
-def returnXml():
+def return_xml():
+    """
+    XML endpoint
+    :return: all catalog information
+    """
     categories = session.query(Category).all()
     root = Element('catalog')
     categories_xml = SubElement(root, 'categories')
     for i in categories:
-        category_xml = xmlCategories(categories_xml,i)
+        category_xml = xml_categories(categories_xml, i)
         items = session.query(Item).filter_by(category_id=i.id).all()
         items_xml = SubElement(category_xml, 'items')
         for j in items:
-            xmlItems(items_xml,j)
+            xml_items(items_xml, j)
 
     return xml_response(tostring(root, encoding="us-ascii", method="xml"),
-                headers={'Content-Type': 'application/xml; charset=utf-8;'})
+                        headers={'Content-Type':
+                        'application/xml; charset=utf-8;'})
 
 
+@csrf.exempt
 @app.route('/catalog/<category_name>.XML')
-def returnXmlCategory(category_name):
+def return_xml_category(category_name):
+    """
+    XML endpoint
+    :param category_name: requested category
+    :return: all catalog information
+    """
     category = session.query(Category).filter_by(name=category_name).all()
     root = Element('catalog')
     for i in category:
-        category_xml = xmlCategories(root,i)
+        category_xml = xml_categories(root, i)
         items = session.query(Item).filter_by(category_id=i.id).all()
         items_xml = SubElement(category_xml, 'items')
         for j in items:
-            xmlItems(items_xml,j)
+            xml_items(items_xml, j)
 
     return xml_response(tostring(root, encoding="us-ascii", method="xml"),
-                headers={'Content-Type': 'application/xml; charset=utf-8;'})
+                        headers={'Content-Type':
+                        'application/xml; charset=utf-8;'})
 
 
 def login_required(func):
@@ -110,14 +135,16 @@ def login_required(func):
     return func_wrapper
 
 
+@csrf.exempt
 @app.route('/login')
-def showLogin():
+def show_login():
         state = ''.join(random.choice(string.ascii_uppercase + string.digits)
                         for x in xrange(32))
         login_session['state'] = state
         return render_template('login.html', STATE=state)
 
 
+@csrf.exempt
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
     if request.args.get('state') != login_session['state']:
@@ -199,6 +226,7 @@ def gconnect():
     return output
 
 
+@csrf.exempt
 @app.route('/gdisconnect')
 def gdisconnect():
     if 'access_token' not in login_session:
@@ -229,30 +257,38 @@ def gdisconnect():
         return response
 
 
+@csrf.exempt
 @app.route('/catalog/<category_name>.JSON')
-def itemsJSON(category_name):
+def items_json(category_name):
+    """
+    Fetches and returns items for requested category
+    :param category_name: requested category
+    :return: data in json format
+    """
     category = session.query(Category).filter_by(name=category_name).one()
     items = session.query(Item).filter_by(
         category_id=category.id).all()
     return jsonify(Items=[i.serialize for i in items])
 
 
+@csrf.exempt
 @app.route('/catalog.JSON')
-def catalogJSON():
+def catalog_json():
     """
-    JSON endpoint
+    Fetches and returns all data in json format
     """
     categories = session.query(Category).all()
     return jsonify(Category=[i.serialize for i in categories])
 
 
+@csrf.exempt
 @app.route('/')
 def catalog():
     """
     catalog returns all categories and lists items ascending by datetime
     """
     categories = session.query(Category).all()
-    items = session.query(Item).order_by(Item.creationDateTime.asc()).\
+    items = session.query(Item).order_by(Item.creationDateTime.desc()).\
         limit(10)
     dict = {}
     for i in categories:
@@ -262,10 +298,11 @@ def catalog():
                            categories=categories, items=items, dict=dict)
 
 
+@csrf.exempt
 @app.route('/catalog/<category_name>/Items')
-def catalogSelected(category_name):
+def catalog_selected(category_name):
     """
-    catalogSelected fetches and returns items from selected category
+    catalog_selected fetches and returns items from selected category
     Args:
         category_name: name of category
     """
@@ -278,6 +315,7 @@ def catalogSelected(category_name):
         items=items, category_name=category_name)
 
 
+@csrf.exempt
 @app.route('/catalog/<category_name>/<item_name>')
 def item(category_name, item_name):
     """
@@ -292,55 +330,60 @@ def item(category_name, item_name):
     return render_template('item.html', category=category, item=item,
                            login_session=login_session)
 
+
 def allowed_file(filename):
     return '.' in filename and filename.rsplit('.', 1)[1] in ALLOWED_EXTENSIONS
 
-def saveFile(file):
+
+def save_file(file):
     filename = secure_filename(file.filename)
-    filename = ''.join([filename.rsplit('.',1)[0],
-                    str(datetime.datetime.now().microsecond),'.',
-                    filename.rsplit('.',1)[1]])
-    os.path.join(app.config['UPLOAD_FOLDER'], filename)
-    file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
+    if len(filename) > 3:
+        filename = ''.join([filename.rsplit('.', 1)[0],
+                            str(datetime.datetime.now().microsecond), '.',
+                            filename.rsplit('.', 1)[1]])
+        os.path.join(app.config['UPLOAD_FOLDER'], filename)
+        file.save(os.path.join(app.config['UPLOAD_FOLDER'], filename))
     return filename
 
-def deleteFile(file):
+
+def delete_file(file):
     try:
         if len(file) > 4:
-            os.remove(os.path.join(app.config['UPLOAD_FOLDER'],file))
+            os.remove(os.path.join(app.config['UPLOAD_FOLDER'], file))
     except IOError as e:
         print "I/O error({0}): {1}".format(e.errno, e.strerror)
 
-def updateFile(old_file, new_file):
-    '''
+
+def update_file(old_file, new_file):
+    """
     :param old_file: path
     :param new_file: object
     :return: new file name
-    '''
-    deleteFile(old_file)
-    return saveFile(new_file)
+    """
+    delete_file(old_file)
+    return save_file(new_file)
 
 
 @login_required
+@csrf.exempt
 @app.route('/catalog/New', methods=['GET', 'POST'])
-def newItem():
+def new_item():
     """
-    newItem adds new item if user is logged in
+    new_item adds new item if user is logged in
     """
     if request.method == 'POST':
         file = request.files['file']
-        if file and allowed_file(file.filename):
-            filename = saveFile(file)
+        if file.filename == "" or (file and allowed_file(file.filename)):
+            filename = save_file(file)
             newItem = Item(name=request.form['item'], description=request.form[
                     'description'], category_id=request.form['category'],
                     image=filename, creationDateTime=datetime.datetime.now())
             session.add(newItem)
             session.commit()
             category_name = session.query(Category).\
-            filter_by(id=request.form['category']).one().name
-
-            return redirect(url_for('catalogSelected',
-                            category_name=category_name, filename=filename))
+                filter_by(id=request.form['category']).one().name
+            return redirect(url_for('catalog_selected',
+                                    category_name=category_name))
     else:
         categories = session.query(Category).all()
         return render_template('add_item.html', login_session=login_session,
@@ -348,11 +391,12 @@ def newItem():
 
 
 @login_required
+@csrf.exempt
 @app.route('/catalog/<category_name>/<item_name>/Edit',
            methods=['GET', 'POST'])
-def editItem(category_name, item_name):
+def edit_item(category_name, item_name):
     """
-    editItem edits item if user is logged in
+    edit_item edits item if user is logged in
     Args:
         category_name: name of category
         item_name: name of item in the category
@@ -369,12 +413,13 @@ def editItem(category_name, item_name):
             editedItem.description = request.form['description']
         if request.form['category']:
             editedItem.category_id = request.form['category']
-        if request.files['file'] and allowed_file(request.files['file'].filename):
-            editedItem.image = updateFile(editedItem.image,
-                                          request.files['file'])
+        if request.files['file'] and allowed_file(
+                request.files['file'].filename):
+            editedItem.image = update_file(editedItem.image,
+                                           request.files['file'])
         session.add(editedItem)
         session.commit()
-        return redirect(url_for('catalogSelected',
+        return redirect(url_for('catalog_selected',
                                 category_name=category_name))
     else:
         return render_template('edit_item.html', category_name=category.name,
@@ -385,9 +430,9 @@ def editItem(category_name, item_name):
 @login_required
 @app.route('/catalog/<category_name>/<item_name>/Delete',
            methods=['GET', 'POST'])
-def deleteItem(category_name, item_name):
+def delete_item(category_name, item_name):
     """
-    deleteItem deletes item if user is logged in
+    delete_item deletes item if user is logged in
     Args:
         category_name: name of category
         item_name: name of item in the category
@@ -396,7 +441,12 @@ def deleteItem(category_name, item_name):
     itemToDelete = session.query(Item).filter_by(name=item_name,
                                                  category_id=category.id).one()
     if request.method == 'POST':
-        deleteFile(itemToDelete.image)
+        token = login_session.pop('_csrf_token', None)
+        if not token or token != request.form['_csrf_token']:
+            response = make_response(json.dumps('CSRF token issue.', 400))
+            response.headers['Content-Type'] = 'application/json'
+            return response
+        delete_file(itemToDelete.image)
         session.delete(itemToDelete)
         session.commit()
         return redirect(url_for('catalog', category_name=category_name))
@@ -405,8 +455,16 @@ def deleteItem(category_name, item_name):
                                login_session=login_session, category=category)
 
 
+def generate_csrf_token():
+    if '_csrf_token' not in login_session:
+        token = "test"
+        login_session['_csrf_token'] = token
+    return login_session['_csrf_token']
+
+
 if __name__ == '__main__':
     app.secret_key = 'bPpAwqouObw5aCWYAhgSRbVn'
     app.config['SESSION_TYPE'] = 'filesystem'
+    app.jinja_env.globals['csrf_token'] = generate_csrf_token
     app.debug = True
     app.run(host='0.0.0.0', port=8080)
